refactor: after review, move function, remove unused export refs: #296
