+++ api/src/core/adapters/hal/getScholarlyArticle.ts
@@ -1,5 +1,5 @@
 import { GetScholarlyArticle } from "../../ports/GetScholarlyArticle";
-import { identifersUtils } from "../../utils";
+import { identifersUtils } from "../../../tools/identifiersTools";
 import { halAPIGateway } from "./HalAPI";
 
 export const getScholarlyArticle: GetScholarlyArticle = async halDocId => {
