+++ web/src/ui/pages/softwareDetails/HeaderDetailCard.tsx
@@ -137,7 +137,10 @@ export const HeaderDetailCard = memo((props: Props) => {
                                                         author.affiliations?.length <=
                                                             0))) && (
                                                 <a
-                                                    href={author.url ?? author?.identifiers?.[0].url?.toString()}
+                                                    href={
+                                                        author.url ??
+                                                        author?.identifiers?.[0]?.url?.toString()
+                                                    }
                                                     className={classes.authorLink}
                                                     key={author.name}
                                                 >
