+++ web/src/ui/shared/LogoURLButton.tsx
@@ -73,9 +73,7 @@ const resolveLogoFromType = (
             };
         case "Orcid":
             return {
-                URLlogo: new URL(
-                    "https://orcid.org/assets/vectors/orcid.logo.icon.svg"
-                ),
+                URLlogo: new URL("https://orcid.org/assets/vectors/orcid.logo.icon.svg"),
                 textFromURL: "ORCID"
             };
         case "wikidata":
